refactor: migrate Button and Carousel components to core directory, improving code organization and maintainability
